refactor: moved product and category into new package files
