+++ src/main/java/com/amalitech/gpuconfigurator/controller/category/CategoryController.java
@@ -4,7 +4,7 @@ package com.amalitech.gpuconfigurator.controller.category;
 import com.amalitech.gpuconfigurator.dto.AllCategoryResponse;
 import com.amalitech.gpuconfigurator.dto.CategoryRequestDto;
 import com.amalitech.gpuconfigurator.model.category.Category;
-import com.amalitech.gpuconfigurator.service.CategoryService;
+import com.amalitech.gpuconfigurator.service.category.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
